+++ dwds/lib/src/debugging/exceptions.dart
@@ -2,18 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.import 'dart:async';
 
-import 'package:shelf/shelf.dart' as shelf;
-
 class ScriptNotFound implements Exception {
   final String scriptPath;
-  final shelf.Response response;
 
-  ScriptNotFound(this.scriptPath, this.response) : super();
+  ScriptNotFound(this.scriptPath) : super();
 
   @override
-  String toString() => '''
-Failed to load script at path: $scriptPath
-
-Response status code: ${response.statusCode}
-''';
+  String toString() => 'Failed to load script at path: $scriptPath';
 }