+++ dwds/test/fixtures/server.dart
@@ -5,12 +5,13 @@
 import 'dart:io';
 
 import 'package:build_daemon/data/build_status.dart' as daemon;
-import 'package:dwds/asset_handler.dart';
 import 'package:dwds/data/build_result.dart';
 import 'package:dwds/dwds.dart';
 import 'package:http_multi_server/http_multi_server.dart';
+import 'package:logging/logging.dart';
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
+import 'package:shelf_proxy/shelf_proxy.dart';
 import 'package:test/test.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
@@ -19,6 +20,7 @@ class TestServer {
   final String target;
   final Dwds dwds;
   final Stream<BuildResult> buildResults;
+  final AssetReader assetReader;
 
   TestServer._(
     this.target,
@@ -26,6 +28,7 @@ class TestServer {
     this.dwds,
     this.buildResults,
     bool autoRun,
+    this.assetReader,
   ) {
     if (autoRun) {
       dwds.connectedApps.listen((connection) {
@@ -72,14 +75,16 @@ class TestServer {
       throw StateError('Unexpected Daemon build result: $result');
     });
 
-    var assetHandler =
-        BuildRunnerAssetHandler(assetServerPort, target, 'localhost', port);
+    var logWriter = (Level level, String message) => printOnFailure(message);
+
+    var assetReader = ProxyServerAssetReader(
+        assetServerPort, target, 'localhost', port, logWriter);
 
     var dwds = await Dwds.start(
-      assetHandler: assetHandler,
+      assetReader: assetReader,
       buildResults: filteredBuildResults,
       chromeConnection: chromeConnection,
-      logWriter: (level, message) => printOnFailure(message),
+      logWriter: logWriter,
       reloadConfiguration: reloadConfiguration,
       serveDevTools: serveDevTools,
       enableDebugExtension: enableDebugExtension,
@@ -94,14 +99,17 @@ class TestServer {
     var server = await HttpMultiServer.bind('localhost', port);
     shelf_io.serveRequests(
         server,
-        pipeline.addMiddleware(dwds.middleware).addHandler(
-            Cascade().add(dwds.handler).add(assetHandler.handler).handler));
+        pipeline.addMiddleware(dwds.middleware).addHandler(Cascade()
+            .add(dwds.handler)
+            .add(proxyHandler('http://localhost:$assetServerPort/$target/'))
+            .handler));
     return TestServer._(
       target,
       server,
       dwds,
       filteredBuildResults,
       autoRun,
+      assetReader,
     );
   }
 }