+++ webdev/lib/src/serve/webdev_server.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:build_daemon/data/build_status.dart' as daemon;
-import 'package:dwds/asset_handler.dart';
 import 'package:dwds/data/build_result.dart';
 import 'package:dwds/dwds.dart';
 import 'package:http_multi_server/http_multi_server.dart';
@@ -94,18 +93,19 @@ class WebDevServer {
       throw StateError('Unexpected Daemon build result: $result');
     });
 
-    Handler assetHandler;
+    var cascade = Cascade();
     Dwds dwds;
     if (options.configuration.enableInjectedClient) {
-      var buildRunnerAssetHandler = BuildRunnerAssetHandler(
+      var assetReader = ProxyServerAssetReader(
         options.daemonPort,
         options.target,
         options.configuration.hostname,
         options.port,
+        logWriter,
       );
       dwds = await Dwds.start(
         hostname: options.configuration.hostname,
-        assetHandler: buildRunnerAssetHandler,
+        assetReader: assetReader,
         buildResults: filteredBuildResults,
         chromeConnection: () async =>
             (await Chrome.connectedInstance).chromeConnection,
@@ -118,14 +118,11 @@ class WebDevServer {
         enableDebugging: options.configuration.debug,
       );
       pipeline = pipeline.addMiddleware(dwds.middleware);
-      assetHandler = Cascade()
-          .add(dwds.handler)
-          .add(buildRunnerAssetHandler.handler)
-          .handler;
-    } else {
-      assetHandler = proxyHandler(
-          'http://localhost:${options.daemonPort}/${options.target}/');
+      cascade = cascade.add(dwds.handler);
     }
+
+    cascade = cascade.add(proxyHandler(
+        'http://localhost:${options.daemonPort}/${options.target}/'));
 
     var hostname = options.configuration.hostname;
     var tlsCertChain = options.configuration.tlsCertChain;
@@ -144,7 +141,7 @@ class WebDevServer {
       server = await HttpMultiServer.bind(hostname, options.port);
     }
 
-    shelf_io.serveRequests(server, pipeline.addHandler(assetHandler));
+    shelf_io.serveRequests(server, pipeline.addHandler(cascade.handler));
     return WebDevServer._(
       options.target,
       server,