+++ dwds/CHANGELOG.md
@@ -5,6 +5,11 @@
 **Breaking Changes:**
 - No longer use the `BuildResult` abstraction from `package:build_daemon` but
   require a similar abstraction provided by this package.
+- `AssetHandler` has been renamed to `AssetReader` and no longer provides a
+  generic resource handler. Specific methods for the required resources are now
+  clearly defined. The new abstraction is now consumed through `dwds.dart`.
+- `BuildRunnerAssetHandler` has been renamed to `ProxyServerAssetReader` and is
+  now consumed through `dwds.dart`.
 
 ## 0.9.0